Added gradients
Sampling two points around the triangle, we can create a gradient and use that for the displayed triangle.
Unfortunately, this significantly increases the size of the output file, reaching about 10MB with the current settings (8192x8192; radius 16-64)

Created ColorMap class which stores the PerlinGen objects and the functionality to sample from them to make an hsl color.